requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
asgiref=3.8.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
bcrypt=4.1.3=pypi_0
boto3=1.34.144=pypi_0
botocore=1.34.144=pypi_0
build=1.2.1=pypi_0
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.7.2=hecd8cb5_0
cachetools=5.3.3=pypi_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
chroma-hnswlib=0.7.5=pypi_0
chromadb=0.5.4=pypi_0
click=8.1.7=pypi_0
cohere=5.5.8=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.1.4=pyhd8ed1ab_0
dataclasses-json=0.6.7=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
deprecated=1.2.14=pypi_0
distro=1.9.0=pypi_0
dnspython=2.6.1=pypi_0
email-validator=2.2.0=pypi_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.111.1=pypi_0
fastapi-cli=0.0.4=pypi_0
fastavro=1.9.5=pypi_0
filelock=3.15.4=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.1=pypi_0
google-auth=2.32.0=pypi_0
googleapis-common-protos=1.63.2=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.64.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.23.4=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib-resources=6.4.0=pypi_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.26.0=pyh3cd1d5f_0
ipython=8.17.2=pyh31c8845_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
jmespath=1.0.1=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.2=py311hecd8cb5_0
kubernetes=30.1.0=pypi_0
langchain=0.2.7=pypi_0
langchain-cohere=0.1.9=pypi_0
langchain-community=0.2.7=pypi_0
langchain-core=0.2.18=pypi_0
langchain-experimental=0.0.62=pypi_0
langchain-openai=0.1.16=pypi_0
langchain-text-splitters=0.2.2=pypi_0
langchainhub=0.1.20=pypi_0
langgraph=0.1.8=pypi_0
langserve=0.2.2=pypi_0
langsmith=0.1.85=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
libsodium=1.0.18=hbcb3906_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.21.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.16.3=pypi_0
openai=1.35.13=pypi_0
openssl=3.0.14=h46256e1_0
opentelemetry-api=1.25.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.25.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.25.0=pypi_0
opentelemetry-instrumentation=0.46b0=pypi_0
opentelemetry-instrumentation-asgi=0.46b0=pypi_0
opentelemetry-instrumentation-fastapi=0.46b0=pypi_0
opentelemetry-proto=1.25.0=pypi_0
opentelemetry-sdk=1.25.0=pypi_0
opentelemetry-semantic-conventions=0.46b0=pypi_0
opentelemetry-util-http=0.46b0=pypi_0
orjson=3.10.6=pypi_0
overrides=7.7.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parameterized=0.9.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pip=24.0=pypi_0
platformdirs=4.0.0=pyhd8ed1ab_0
posthog=3.5.0=pypi_0
prompt-toolkit=3.0.41=pyha770c72_0
prompt_toolkit=3.0.41=hd8ed1ab_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pypdf=4.3.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.1.0=pypi_0
pyproject-toml=0.0.10=pypi_0
python=3.11.9=hf27a42d_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
readline=8.2=hca72f7f_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.19.0=pypi_0
rsa=4.9=pypi_0
s3transfer=0.10.2=pypi_0
setuptools=69.5.1=pypi_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.31=pypi_0
sqlite=3.41.2=h6c40b1e_0
sse-starlette=1.8.2=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.37.2=pypi_0
sympy=1.13.0=pypi_0
tabulate=0.9.0=pypi_0
tavily-python=0.3.4=pypi_0
tenacity=8.5.0=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.12=h5d9f67b_0
tokenizers=0.19.1=pypi_0
toml=0.10.2=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.13.0=pyhd8ed1ab_0
typer=0.12.3=pypi_0
types-requests=2.32.0.20240712=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.1=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.22.0=pypi_0
wcwidth=0.2.10=pyhd8ed1ab_0
websocket-client=1.8.0=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=pypi_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h6c40b1e_1
yarl=1.9.4=pypi_0
zeromq=4.3.5=hcec6c5f_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h4dc903c_0
